Add files via upload
The objective was to apply number of different learnings gained throughout the Cybersecurity Bootcamp and to take the role of a penetration tester and SOC analyst.  Implement alarms and thresholds in Elasticsearch Watcher. Find vulnerabilities / exploits to gain access to a vulnerable WordPress site / server. Use Wireshark to analyze and find malicious traffic on the wire. Create a team presentation with the assessment, analysis and hardening of the vulnerable system. Read through the Red Team, Blue Team and Network Analysis files to the see the work, process and findings that lead up to creation of the final presentation.  View the final Google slide project presentation in this repository. Note: This presentation was created as a group with classmates and focuses on the offensive side of the project.